Add files via upload
The ble_mouse.ino code was updated to fix issues with the x-axis movement of the joystick being inverted, and mouse movement being produced when the joystick was not being used. To resolve these issues the joystick sensibility in the x-axis was inverted and a dead zone was applied to the joystick. The application of the dead zone also resolved some issues with jittering of the mouse.